+++ client/apps/eternum-mobile/src/pages/realm/ui/components/production-widgets-section.tsx
